+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG.md
 
+* build3, stable, latest
+    * support UTF-8 quoted-printable
+
 * build2, stable, latest
     * support UTF-8 8bit mail